ISSUE #707 replaced long long with int64_t and made a few other tweaks for gcc

+++ bouncer/src/repo/core/model/bson/repo_bson.cpp
@@ -425,7 +425,7 @@ double RepoBSON::getDoubleField(const std::string &label) const
 	}
 }
 
-long long RepoBSON::getLongField(const std::string& label) const
+int64_t RepoBSON::getLongField(const std::string& label) const
 {
 	try
 	{
