+++ bouncer/src/repo/manipulator/modelconvertor/import/repo_model_import_synchro.cpp
@@ -135,7 +135,7 @@ repo::core::model::MetadataNode* SynchroModelImport::createMetaNode(
 		//Check if it is a number, if it is, store it as a number
 
 		try {
-			v = boost::lexical_cast<long long>(value);
+			v = boost::lexical_cast<int64_t>(value);
 		}
 		catch (boost::bad_lexical_cast&)
 		{
